% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/websocket_addin.R
\name{auto_capture_error}
\alias{auto_capture_error}
\title{Auto-capture error context and show notification}
\usage{
auto_capture_error()
}
\description{
Auto-capture error context and show notification
}
