% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_agent.R
\name{calculate_automatic_effect_sizes}
\alias{calculate_automatic_effect_sizes}
\title{Calculate Automatic Effect Sizes}
\usage{
calculate_automatic_effect_sizes(df, variables)
}
\arguments{
\item{df}{Dataframe}

\item{variables}{Variable specifications}
}
\value{
Automatic effect size results
}
\description{
Calculate Automatic Effect Sizes
}
