% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_agent.R
\name{calculate_cohens_d}
\alias{calculate_cohens_d}
\title{Calculate Cohen's d Effect Size}
\usage{
calculate_cohens_d(group1_data, group2_data)
}
\arguments{
\item{group1_data}{First group data}

\item{group2_data}{Second group data}
}
\value{
Effect size results
}
\description{
Calculate Cohen's d Effect Size
}
