% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_agent.R
\name{calculate_cohens_d_analysis}
\alias{calculate_cohens_d_analysis}
\title{Calculate Cohen's d Analysis}
\usage{
calculate_cohens_d_analysis(df, variables)
}
\arguments{
\item{df}{Dataframe}

\item{variables}{Variable specifications}
}
\value{
Cohen's d results
}
\description{
Calculate Cohen's d Analysis
}
