% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_agent.R
\name{calculate_eta_squared_analysis}
\alias{calculate_eta_squared_analysis}
\title{Calculate Eta-squared Analysis}
\usage{
calculate_eta_squared_analysis(df, variables)
}
\arguments{
\item{df}{Dataframe}

\item{variables}{Variable specifications}
}
\value{
Eta-squared results
}
\description{
Calculate Eta-squared Analysis
}
