% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_agent.R
\name{calculate_odds_ratio_analysis}
\alias{calculate_odds_ratio_analysis}
\title{Calculate Odds Ratio Analysis}
\usage{
calculate_odds_ratio_analysis(df, variables)
}
\arguments{
\item{df}{Dataframe}

\item{variables}{Variable specifications}
}
\value{
Odds ratio results
}
\description{
Calculate Odds Ratio Analysis
}
