% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_agent.R
\name{check_test_assumptions}
\alias{check_test_assumptions}
\title{Check Test Assumptions}
\usage{
check_test_assumptions(group1_data, group2_data)
}
\arguments{
\item{group1_data}{First group data}

\item{group2_data}{Second group data}
}
\value{
Assumption check results
}
\description{
Check Test Assumptions
}
