% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_agent.R
\name{convert_tables_to_serializable}
\alias{convert_tables_to_serializable}
\title{Convert table objects to serializable format}
\usage{
convert_tables_to_serializable(obj)
}
\arguments{
\item{obj}{Object to check for table objects}
}
\value{
Object with table objects converted
}
\description{
Convert table objects to serializable format
}
