% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_agent.R
\name{create_modeling_workflow}
\alias{create_modeling_workflow}
\title{Create Modeling Workflow Steps}
\usage{
create_modeling_workflow(
  algorithms,
  options,
  dataframe_name,
  target_variable_name,
  custom_inputs = NULL
)
}
\arguments{
\item{algorithms}{Selected algorithms}

\item{options}{Additional options}

\item{dataframe_name}{Name of the dataframe}

\item{target_variable_name}{Name of the target variable}
}
\value{
List of workflow steps
}
\description{
Create Modeling Workflow Steps
}
