% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_agent.R
\name{create_statistical_workflow}
\alias{create_statistical_workflow}
\title{Create Statistical Workflow Steps}
\usage{
create_statistical_workflow(
  analysis_options,
  variables,
  method_options,
  custom_inputs = NULL
)
}
\arguments{
\item{analysis_options}{Selected analysis types}

\item{variables}{Variable specifications}

\item{method_options}{Method preferences}

\item{custom_inputs}{Custom analysis specifications}
}
\value{
List of workflow steps
}
\description{
Create Statistical Workflow Steps
}
