% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_agent.R
\name{create_visualization_workflow}
\alias{create_visualization_workflow}
\title{Create visualization workflow steps}
\usage{
create_visualization_workflow(plot_types, variables, data_info)
}
\arguments{
\item{plot_types}{List of selected plot types}

\item{variables}{List of selected variables}

\item{data_info}{Data structure information}
}
\value{
List of workflow steps
}
\description{
Create visualization workflow steps
}
