% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_agent.R
\name{create_visualization_workflow_from_plots}
\alias{create_visualization_workflow_from_plots}
\title{Create visualization workflow from individual plot configurations}
\usage{
create_visualization_workflow_from_plots(plots, data_info)
}
\arguments{
\item{plots}{List of configured plots}

\item{data_info}{Data structure information}
}
\value{
List of workflow steps
}
\description{
Create visualization workflow from individual plot configurations
}
