% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_agent.R
\name{execute_categorical_tests}
\alias{execute_categorical_tests}
\title{Execute Categorical Tests}
\usage{
execute_categorical_tests(df, variables, method_options)
}
\arguments{
\item{df}{Dataframe to analyze}

\item{variables}{Variable specifications}

\item{method_options}{Method preferences}
}
\value{
Categorical test results
}
\description{
Execute Categorical Tests
}
