% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_code_functions.R
\name{execute_code}
\alias{execute_code}
\title{Execute code directly in R console}
\usage{
execute_code(code)
}
\arguments{
\item{code}{The code to execute}
}
\description{
Execute code directly in R console
}
