% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_agent.R
\name{execute_custom_modeling}
\alias{execute_custom_modeling}
\title{Execute Custom Modeling}
\usage{
execute_custom_modeling(
  df,
  target_variable,
  custom_inputs,
  selected_variables = NULL
)
}
\arguments{
\item{df}{Dataframe to analyze}

\item{target_variable}{Target variable}

\item{custom_inputs}{Custom modeling specifications}
}
\value{
Custom modeling results
}
\description{
Execute Custom Modeling
}
