% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_agent.R
\name{execute_custom_statistical_analysis}
\alias{execute_custom_statistical_analysis}
\title{Execute Custom Statistical Analysis}
\usage{
execute_custom_statistical_analysis(
  df,
  variables,
  method_options,
  custom_inputs
)
}
\arguments{
\item{df}{Dataframe to analyze}

\item{variables}{Variable specifications}

\item{method_options}{Method preferences}

\item{custom_inputs}{Custom analysis specifications}
}
\value{
Custom analysis results
}
\description{
Execute Custom Statistical Analysis
}
