% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_agent.R
\name{execute_dimensionality_reduction}
\alias{execute_dimensionality_reduction}
\title{Execute Dimensionality Reduction}
\usage{
execute_dimensionality_reduction(df, options = NULL, selected_variables = NULL)
}
\arguments{
\item{df}{Dataframe to analyze}

\item{options}{Additional options}

\item{selected_variables}{Variables selected for this algorithm}
}
\value{
Dimensionality reduction results
}
\description{
Execute Dimensionality Reduction
}
