% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_agent.R
\name{execute_effect_size_analysis}
\alias{execute_effect_size_analysis}
\title{Execute Effect Size Analysis}
\usage{
execute_effect_size_analysis(df, variables, method_options)
}
\arguments{
\item{df}{Dataframe to analyze}

\item{variables}{Variable specifications}

\item{method_options}{Method preferences}
}
\value{
Effect size analysis results
}
\description{
Execute Effect Size Analysis
}
