% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_agent.R
\name{execute_multiple_testing_correction}
\alias{execute_multiple_testing_correction}
\title{Execute Multiple Testing Correction}
\usage{
execute_multiple_testing_correction(df, variables)
}
\arguments{
\item{df}{Dataframe to analyze}

\item{variables}{Variable specifications}
}
\value{
Multiple testing correction results
}
\description{
Execute Multiple Testing Correction
}
