% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_agent.R
\name{execute_nonparametric_tests}
\alias{execute_nonparametric_tests}
\title{Execute Nonparametric Tests}
\usage{
execute_nonparametric_tests(df, variables, method_options)
}
\arguments{
\item{df}{Dataframe to analyze}

\item{variables}{Variable specifications}

\item{method_options}{Method preferences}
}
\value{
Nonparametric test results
}
\description{
Execute Nonparametric Tests
}
