% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/websocket_addin.R
\name{execute_plot_analysis}
\alias{execute_plot_analysis}
\title{Execute analysis commands and collect results}
\usage{
execute_plot_analysis(commands)
}
\description{
Execute analysis commands and collect results
}
