% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_agent.R
\name{execute_plot_code}
\alias{execute_plot_code}
\title{Execute plot code and capture results}
\usage{
execute_plot_code(plot_code, dataframe, variables)
}
\arguments{
\item{plot_code}{R code to execute}

\item{dataframe}{Name of the dataframe}

\item{variables}{List of selected variables}
}
\value{
List containing execution results
}
\description{
Execute plot code and capture results
}
