% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_agent.R
\name{execute_power_analysis}
\alias{execute_power_analysis}
\title{Execute Power Analysis}
\usage{
execute_power_analysis(df, variables, method_options)
}
\arguments{
\item{df}{Dataframe to analyze}

\item{variables}{Variable specifications}

\item{method_options}{Method preferences}
}
\value{
Power analysis results
}
\description{
Execute Power Analysis
}
