% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_code_functions.R
\name{execute_r_function}
\alias{execute_r_function}
\title{Execute R function directly from frontend}
\usage{
execute_r_function(function_name, param1, param2 = NULL)
}
\arguments{
\item{function_name}{Name of the R function to call}

\item{param1}{First parameter}

\item{param2}{Second parameter}
}
\description{
Execute R function directly from frontend
}
