% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_agent.R
\name{execute_results_compilation}
\alias{execute_results_compilation}
\title{Execute Results Compilation}
\usage{
execute_results_compilation(df, target_variable, algorithms)
}
\arguments{
\item{df}{Dataframe to analyze}

\item{target_variable}{Target variable}

\item{algorithms}{Selected algorithms}
}
\value{
Results compilation
}
\description{
Execute Results Compilation
}
