% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_agent.R
\name{execute_visualization_step}
\alias{execute_visualization_step}
\title{Execute a visualization step}
\usage{
execute_visualization_step(step, dataframe, variables, options)
}
\arguments{
\item{step}{Step number}

\item{dataframe}{Name of the dataframe}

\item{variables}{List of selected variables (legacy parameter)}

\item{options}{List of visualization options including configured plots}
}
\value{
List containing execution results
}
\description{
Execute a visualization step
}
