% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/websocket_addin.R
\name{extract_plot_data}
\alias{extract_plot_data}
\title{Extract data variables from plot command}
\usage{
extract_plot_data(command, plot_type)
}
\description{
Extract data variables from plot command
}
