% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/websocket_addin.R
\name{find_exact_matches}
\alias{find_exact_matches}
\title{Find exact matches for keywords in object names}
\usage{
find_exact_matches(keywords, objects)
}
\description{
Find exact matches for keywords in object names
}
