% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/websocket_addin.R
\name{find_last_plot_command}
\alias{find_last_plot_command}
\title{Update workspace index
Find the last plot command in R history}
\usage{
find_last_plot_command()
}
\description{
Update workspace index
Find the last plot command in R history
}
