% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/websocket_addin.R
\name{find_partial_matches}
\alias{find_partial_matches}
\title{Find partial and fuzzy matches for keywords in object names}
\usage{
find_partial_matches(keywords, objects)
}
\description{
Find partial and fuzzy matches for keywords in object names
}
