% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/websocket_addin.R
\name{generate_analysis_commands}
\alias{generate_analysis_commands}
\title{Generate analysis commands based on plot type}
\usage{
generate_analysis_commands(plot_type, data_var)
}
\description{
Generate analysis commands based on plot type
}
