% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_agent.R
\name{generate_bar_code}
\alias{generate_bar_code}
\title{Generate bar chart code}
\usage{
generate_bar_code(dataframe, variables, options)
}
\arguments{
\item{dataframe}{Name of the dataframe}

\item{variables}{List of selected variables}

\item{options}{List of visualization options}
}
\value{
R code string for bar chart
}
\description{
Generate bar chart code
}
