% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_agent.R
\name{generate_boxplot_code}
\alias{generate_boxplot_code}
\title{Generate boxplot code}
\usage{
generate_boxplot_code(dataframe, variables, options)
}
\arguments{
\item{dataframe}{Name of the dataframe}

\item{variables}{List of selected variables}

\item{options}{List of visualization options}
}
\value{
R code string for boxplot
}
\description{
Generate boxplot code
}
