% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_agent.R
\name{generate_correlation_code}
\alias{generate_correlation_code}
\title{Generate correlation plot code}
\usage{
generate_correlation_code(dataframe, variables, options)
}
\arguments{
\item{dataframe}{Name of the dataframe}

\item{variables}{List of selected variables}

\item{options}{List of visualization options}
}
\value{
R code string for correlation plot
}
\description{
Generate correlation plot code
}
