% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_agent.R
\name{generate_density_code}
\alias{generate_density_code}
\title{Generate density plot code}
\usage{
generate_density_code(dataframe, variables, options)
}
\arguments{
\item{dataframe}{Name of the dataframe}

\item{variables}{List of selected variables}

\item{options}{List of visualization options}
}
\value{
R code string for density plot
}
\description{
Generate density plot code
}
