% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_agent.R
\name{generate_histogram_code}
\alias{generate_histogram_code}
\title{Generate histogram code}
\usage{
generate_histogram_code(dataframe, variables, options)
}
\arguments{
\item{dataframe}{Name of the dataframe}

\item{variables}{List of selected variables}

\item{options}{List of visualization options}
}
\value{
R code string for histogram
}
\description{
Generate histogram code
}
