% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_agent.R
\name{generate_line_code}
\alias{generate_line_code}
\title{Generate line plot code}
\usage{
generate_line_code(dataframe, variables, options)
}
\arguments{
\item{dataframe}{Name of the dataframe}

\item{variables}{List of selected variables}

\item{options}{List of visualization options}
}
\value{
R code string for line plot
}
\description{
Generate line plot code
}
