% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{generate_na_handling_code}
\alias{generate_na_handling_code}
\title{Generate NA Handling Code}
\usage{
generate_na_handling_code(dataframe, method, df_info)
}
\arguments{
\item{dataframe}{Name of the DataFrame}

\item{method}{Method for handling NAs}

\item{df_info}{DataFrame information}
}
\value{
Character string with R code
}
\description{
Generate NA Handling Code
}
