% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_agent.R
\name{generate_pie_code}
\alias{generate_pie_code}
\title{Generate pie chart code}
\usage{
generate_pie_code(dataframe, variables, options)
}
\arguments{
\item{dataframe}{Name of the dataframe}

\item{variables}{List of selected variables}

\item{options}{List of visualization options}
}
\value{
R code string for pie chart
}
\description{
Generate pie chart code
}
