% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_agent.R
\name{generate_plot_code}
\alias{generate_plot_code}
\title{Generate R code for a specific plot type}
\usage{
generate_plot_code(plot_type, dataframe, variables, options)
}
\arguments{
\item{plot_type}{Type of plot to generate}

\item{dataframe}{Name of the dataframe}

\item{variables}{List of selected variables}

\item{options}{List of visualization options}
}
\value{
R code string for the plot
}
\description{
Generate R code for a specific plot type
}
