% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_agent.R
\name{generate_plot_code_by_type}
\alias{generate_plot_code_by_type}
\title{Generate plot code by type with individual configuration}
\usage{
generate_plot_code_by_type(plot_type, plot_config, dataframe, options)
}
\arguments{
\item{plot_type}{Type of plot to generate}

\item{plot_config}{Plot configuration object}

\item{dataframe}{Name of the dataframe}

\item{options}{List of visualization options}
}
\value{
R code string for the plot
}
\description{
Generate plot code by type with individual configuration
}
