% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_agent.R
\name{generate_plot_code_from_config}
\alias{generate_plot_code_from_config}
\title{Generate R code from plot configuration}
\usage{
generate_plot_code_from_config(plot_config, dataframe, options)
}
\arguments{
\item{plot_config}{Plot configuration object}

\item{dataframe}{Name of the dataframe}

\item{options}{List of visualization options}
}
\value{
R code string for the plot
}
\description{
Generate R code from plot configuration
}
