% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_agent.R
\name{generate_qqplot_code}
\alias{generate_qqplot_code}
\title{Generate Q-Q plot code}
\usage{
generate_qqplot_code(dataframe, variables, options)
}
\arguments{
\item{dataframe}{Name of the dataframe}

\item{variables}{List of selected variables}

\item{options}{List of visualization options}
}
\value{
R code string for Q-Q plot
}
\description{
Generate Q-Q plot code
}
