% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_agent.R
\name{generate_residual_code}
\alias{generate_residual_code}
\title{Generate residual plot code}
\usage{
generate_residual_code(dataframe, variables, options)
}
\arguments{
\item{dataframe}{Name of the dataframe}

\item{variables}{List of selected variables}

\item{options}{List of visualization options}
}
\value{
R code string for residual plot
}
\description{
Generate residual plot code
}
