% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_agent.R
\name{generate_trend_code}
\alias{generate_trend_code}
\title{Generate trend analysis code}
\usage{
generate_trend_code(dataframe, variables, options)
}
\arguments{
\item{dataframe}{Name of the dataframe}

\item{variables}{List of selected variables}

\item{options}{List of visualization options}
}
\value{
R code string for trend analysis
}
\description{
Generate trend analysis code
}
