% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{generate_type_conversion_code}
\alias{generate_type_conversion_code}
\title{Generate Data Type Conversion Code}
\usage{
generate_type_conversion_code(dataframe, df_info)
}
\arguments{
\item{dataframe}{Name of the DataFrame}

\item{df_info}{DataFrame information}
}
\value{
Character string with R code
}
\description{
Generate Data Type Conversion Code
}
