% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_agent.R
\name{generate_violin_code}
\alias{generate_violin_code}
\title{Generate violin plot code}
\usage{
generate_violin_code(dataframe, variables, options)
}
\arguments{
\item{dataframe}{Name of the dataframe}

\item{variables}{List of selected variables}

\item{options}{List of visualization options}
}
\value{
R code string for violin plot
}
\description{
Generate violin plot code
}
