% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_agent.R
\name{generate_visualization_step_code}
\alias{generate_visualization_step_code}
\title{Generate visualization step code (for execution engine)}
\usage{
generate_visualization_step_code(step, dataframe, variables, options)
}
\arguments{
\item{step}{Step number}

\item{dataframe}{Name of the dataframe}

\item{variables}{List of selected variables (legacy parameter)}

\item{options}{List of visualization options including configured plots}
}
\value{
R code string for the visualization step
}
\description{
Generate visualization step code (for execution engine)
}
