% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{get_available_dataframes}
\alias{get_available_dataframes}
\title{Get Available DataFrames in Global Environment}
\usage{
get_available_dataframes()
}
\value{
Character vector of DataFrame names
}
\description{
Get Available DataFrames in Global Environment
}
