% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/websocket_addin.R
\name{get_conversation_history}
\alias{get_conversation_history}
\title{Get conversation history as formatted string}
\usage{
get_conversation_history()
}
\description{
Get conversation history as formatted string
}
