% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{get_dataframe_info}
\alias{get_dataframe_info}
\title{Get Basic Information About a DataFrame}
\usage{
get_dataframe_info(dataframe_name)
}
\arguments{
\item{df_name}{Name of the DataFrame}
}
\value{
List with DataFrame information
}
\description{
Get Basic Information About a DataFrame
}
