% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/websocket_addin.R
\name{get_last_executed_code}
\alias{get_last_executed_code}
\title{Get the last executed code from R history}
\usage{
get_last_executed_code()
}
\description{
Get the last executed code from R history
}
