% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_agent.R
\name{get_next_modeling_step}
\alias{get_next_modeling_step}
\title{Get Next Modeling Step Code}
\usage{
get_next_modeling_step(
  dataframe_name,
  step_info,
  target_variable,
  algorithms,
  selected_variables = NULL
)
}
\arguments{
\item{dataframe_name}{Name of the dataframe}

\item{step_info}{Information about the current step}

\item{target_variable}{Target variable}

\item{algorithms}{Selected algorithms}
}
\value{
Next step code to execute
}
\description{
Get Next Modeling Step Code
}
