% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_agent.R
\name{get_next_statistical_step}
\alias{get_next_statistical_step}
\title{Get Next Statistical Analysis Step}
\usage{
get_next_statistical_step(dataframe_name, step_info, variables, method_options)
}
\arguments{
\item{dataframe_name}{Name of the dataframe}

\item{step_info}{Information about the current step}

\item{variables}{Variable specifications}

\item{method_options}{Method preferences}
}
\value{
Next step information
}
\description{
Get Next Statistical Analysis Step
}
