% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_agent.R
\name{get_next_visualization_step}
\alias{get_next_visualization_step}
\title{Get the next visualization step}
\usage{
get_next_visualization_step(step, dataframe, variables, options)
}
\arguments{
\item{step}{Current step number}

\item{dataframe}{Name of the dataframe}

\item{variables}{List of selected variables (legacy parameter)}

\item{options}{List of visualization options including configured plots}
}
\value{
List containing step information and code
}
\description{
Get the next visualization step
}
