% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_agent.R
\name{get_plot_description}
\alias{get_plot_description}
\title{Get description for a plot type}
\usage{
get_plot_description(plot_type, variables)
}
\arguments{
\item{plot_type}{Type of plot}

\item{variables}{List of selected variables}
}
\value{
Description string
}
\description{
Get description for a plot type
}
