% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_agent.R
\name{get_plot_description_from_config}
\alias{get_plot_description_from_config}
\title{Get plot description from configuration}
\usage{
get_plot_description_from_config(plot_config)
}
\arguments{
\item{plot_config}{Plot configuration object}
}
\value{
Description string
}
\description{
Get plot description from configuration
}
