% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/websocket_addin.R
\name{get_relevant_context_for_error}
\alias{get_relevant_context_for_error}
\title{Get relevant context from index based on error type}
\usage{
get_relevant_context_for_error(error_code, error_message)
}
\description{
Get relevant context from index based on error type
}
