% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/websocket_addin.R
\name{get_rstudio_theme}
\alias{get_rstudio_theme}
\title{Get RStudio theme information}
\usage{
get_rstudio_theme()
}
\value{
List containing theme information
}
\description{
Get RStudio theme information
}
