% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rgent.R
\name{insert_code}
\alias{insert_code}
\title{Insert code into active document}
\usage{
insert_code(code = NULL)
}
\arguments{
\item{code}{The code to insert (if NULL, reads from clipboard)}
}
\description{
Insert code into active document
}
