% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_agent.R
\name{is_plot_appropriate}
\alias{is_plot_appropriate}
\title{Check if a plot type is appropriate for the given variables and data}
\usage{
is_plot_appropriate(plot_type, variables, data_info)

is_plot_appropriate(plot_type, variables, data_info)
}
\arguments{
\item{plot_type}{Type of plot to check}

\item{variables}{Selected variables}

\item{data_info}{Data structure information}
}
\value{
Logical indicating if plot is appropriate

Boolean indicating if plot is appropriate
}
\description{
Check if a plot type is appropriate for the given variables and data

Check if a plot type is appropriate for the given data
}
