% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/websocket_addin.R
\name{launch_html_interface}
\alias{launch_html_interface}
\title{Launch HTML interface with HTTPS support for clipboard API}
\usage{
launch_html_interface()
}
\description{
Launch HTML interface with HTTPS support for clipboard API
}
