% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution_engine.R
\name{load_execution_settings}
\alias{load_execution_settings}
\title{Load execution settings}
\usage{
load_execution_settings()
}
\value{
A list containing execution settings
}
\description{
Load execution settings
}
