% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_agent.R
\name{perform_anova_test}
\alias{perform_anova_test}
\title{Perform ANOVA Test for Multiple Groups}
\usage{
perform_anova_test(df, continuous_var, grouping_var, method_options)
}
\arguments{
\item{df}{Dataframe}

\item{continuous_var}{Continuous variable name}

\item{grouping_var}{Grouping variable name}

\item{method_options}{Method preferences}
}
\value{
ANOVA test results
}
\description{
Perform ANOVA Test for Multiple Groups
}
