% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_agent.R
\name{perform_mann_whitney_test}
\alias{perform_mann_whitney_test}
\title{Perform Mann-Whitney U Test}
\usage{
perform_mann_whitney_test(group1_data, group2_data)
}
\arguments{
\item{group1_data}{First group data}

\item{group2_data}{Second group data}
}
\value{
Mann-Whitney test results
}
\description{
Perform Mann-Whitney U Test
}
