% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_agent.R
\name{perform_t_test}
\alias{perform_t_test}
\title{Perform T-Test}
\usage{
perform_t_test(group1_data, group2_data, equal_variance = FALSE)
}
\arguments{
\item{group1_data}{First group data}

\item{group2_data}{Second group data}

\item{equal_variance}{Whether to assume equal variances}
}
\value{
T-test results
}
\description{
Perform T-Test
}
