% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_agent.R
\name{perform_two_group_test}
\alias{perform_two_group_test}
\title{Perform Two Group Test (T-test or Z-test)}
\usage{
perform_two_group_test(df, continuous_var, grouping_var, method_options)
}
\arguments{
\item{df}{Dataframe}

\item{continuous_var}{Continuous variable name}

\item{grouping_var}{Grouping variable name}

\item{method_options}{Method preferences}
}
\value{
Test results
}
\description{
Perform Two Group Test (T-test or Z-test)
}
