% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_agent.R
\name{perform_z_test}
\alias{perform_z_test}
\title{Perform Z-Test}
\usage{
perform_z_test(group1_data, group2_data)
}
\arguments{
\item{group1_data}{First group data}

\item{group2_data}{Second group data}
}
\value{
Z-test results
}
\description{
Perform Z-Test
}
