% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/websocket_addin.R
\name{reconstruct_base_plot_command}
\alias{reconstruct_base_plot_command}
\title{Reconstruct multi-line base R plot command (plot, hist, boxplot, barplot)}
\usage{
reconstruct_base_plot_command(history_lines, start_line, plot_cmd)
}
\description{
Reconstruct multi-line base R plot command (plot, hist, boxplot, barplot)
}
