% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/websocket_addin.R
\name{reconstruct_ggplot_command}
\alias{reconstruct_ggplot_command}
\title{Reconstruct multi-line ggplot2 command}
\usage{
reconstruct_ggplot_command(history_lines, start_line)
}
\description{
Reconstruct multi-line ggplot2 command
}
