% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/websocket_addin.R
\name{reconstruct_qq_command}
\alias{reconstruct_qq_command}
\title{Reconstruct base R QQ plot command possibly spanning multiple lines}
\usage{
reconstruct_qq_command(history_lines, start_line)
}
\description{
Reconstruct base R QQ plot command possibly spanning multiple lines
}
