% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution_engine.R
\name{save_execution_settings}
\alias{save_execution_settings}
\title{Save execution settings}
\usage{
save_execution_settings(settings)
}
\arguments{
\item{settings}{A list containing execution settings}
}
\value{
Invisible NULL
}
\description{
Save execution settings
}
