% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/websocket_addin.R
\name{smart_filter_objects}
\alias{smart_filter_objects}
\title{Smart filtering functions for workspace objects}
\usage{
smart_filter_objects(query, objects)
}
\description{
Smart filtering functions for workspace objects
}
