% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{start_cleaning_agent}
\alias{start_cleaning_agent}
\title{Start Cleaning Agent Workflow}
\usage{
start_cleaning_agent(
  dataframe,
  na_handling = "median",
  cleaning_options = NULL,
  method_options = NULL,
  custom_inputs = NULL,
  other_operations = NULL,
  selected_variables = NULL
)
}
\arguments{
\item{dataframe}{Name of the DataFrame to clean}

\item{na_handling}{Method for handling NAs}

\item{cleaning_options}{List of cleaning operations to perform}

\item{method_options}{List of method choices for each operation}

\item{custom_inputs}{List of custom text inputs for "Other" methods}

\item{other_operations}{Custom cleaning operations specified by user}
}
\value{
List with workflow information
}
\description{
Start Cleaning Agent Workflow
}
