% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_agent.R
\name{start_visualization_agent}
\alias{start_visualization_agent}
\title{Start the visualization agent workflow}
\usage{
start_visualization_agent(dataframe, options)
}
\arguments{
\item{dataframe}{Name of the dataframe to visualize}

\item{options}{List of visualization options including configured plots}
}
\value{
List containing workflow configuration
}
\description{
Start the visualization agent workflow
}
