% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/websocket_addin.R
\name{start_websocket_server}
\alias{start_websocket_server}
\title{Start WebSocket server for R-JavaScript communication}
\usage{
start_websocket_server()
}
\description{
Start WebSocket server for R-JavaScript communication
}
