% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/websocket_addin.R
\name{update_workspace_index_incremental}
\alias{update_workspace_index_incremental}
\title{Incremental workspace update functions}
\usage{
update_workspace_index_incremental()
}
\description{
Incremental workspace update functions
}
